requirements.txt
requests
pillow
numpy
torch
